Move BulletMover in an entity folder.
