+++ src/entities/BulletMover.cpp
@@ -1,4 +1,4 @@
-#include "components/BulletMover.h"
+#include "entities/BulletMover.h"
 
 #include <bullet/btBulletCollisionCommon.h>
 #include <bullet/btBulletDynamicsCommon.h>
